requirements.txt
PyQt5>=5.15
cantools>=38.0.2
python-can>=4.2.0
PyYAML>=6.0
pyqtgraph>=0.13
